+++ src/TitleReceive/TitleReceiveContainer.js
@@ -73,10 +73,12 @@ function TitleReceiveContainer({ history, location, match, mutator, resources })
   useEffect(
     () => {
       if (poLineId) {
+        const filterQuery = `titleId=${titleId} and poLineId==${poLineId} and receivingStatus==${PIECE_STATUS.expected}`;
+
         mutator.pieces.GET({
           params: {
             limit: `${LIMIT_MAX}`,
-            query: `poLineId==${poLineId} and receivingStatus==${PIECE_STATUS.expected} sortby locationId`,
+            query: `${filterQuery} sortby locationId`,
           },
         })
           .then(piecesResponse => getHydratedPieces(piecesResponse, mutator.requests, mutator.items))
@@ -84,7 +86,7 @@ function TitleReceiveContainer({ history, location, match, mutator, resources })
       }
     },
     // eslint-disable-next-line react-hooks/exhaustive-deps
-    [poLineId],
+    [titleId, poLineId],
   );
 
   useEffect(