+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change history for ui-receiving
 
 ## (IN PROGRESS)
+* Also support `circulation` `11.0`. Refs UIPCIR-21.
 
 ## [1.3.3](https://github.com/folio-org/ui-receiving/tree/v1.3.3) (2021-04-21)
 [Full Changelog](https://github.com/folio-org/ui-receiving/compare/v1.3.2...v1.3.3)